Add files via upload
There is also .txt file in my this repository , This is an updated version of that with some few changes.  This script works on dictionary method to save data in terminal/console. 

Working Description :

It has feature to "Add Contact" which ask user to enter their name, and their phone number which is  added to computer memory as dictionary. There is another feature which is "Update contact", it ask user to enter name so that it can search that particular contact, after that it ask new number to replace the existing number. Furthermore, there is "Delete contact" option, like wise "Update contact" it also ask user to enter name so that it can search that particular contact, then it delete the existing contact. Next, it has "View Contact" option to see all  the recently added contact , Unlike recent options it doesn't ask  user to input anything and lastly, this program include an option to close program which is "Exit". It break my loop to close my program.

I had too much fun to create this program. 
Hello, World!